fix(apps/spu-ui): moved upload trigger to setup 1 section and fixed form validation.
