+++ .changeset/all-keys-stick.md
@@ -1,0 +1,5 @@
+---
+"@sophys-web/spu-ui": minor
+---
+
+Moved "upload csv" trigger to setup 1 section.
